fix product list

+++ lib/services/product_services.dart
@@ -19,8 +19,8 @@ class ProductServices {
           title: product['title'],
           description: product['description'],
           price: product['price'],
-          discountPercentage: product['discountPercentage'],
-          rating: product['rating'],
+          discountPercentage: product['discountPercentage'].toDouble(),
+          rating: product['rating'].toDouble(),
           stock: product['stock'],
           brand: product['brand'],
           category: product['category'],
@@ -30,30 +30,7 @@ class ProductServices {
       }).toList();
 
 
-      return productList;
-
-
-      // for (var product in prods) {
-      //   print('Product ID: ${product['id']}');
-      //   print('Product Title: ${product['title']}');
-      //   print('Product Description: ${product['description']}');
-      //   print('Product Price: ${product['price']}');
-      //   print('Product Discount Percentage: ${product['discountPercentage']}');
-      //   print('Product Rating: ${product['rating']}');
-      //   print('Product Stock: ${product['stock']}');
-      //   print('Product Brand: ${product['brand']}');
-      //   print('Product Category: ${product['category']}');
-      //   print('Product Thumbnail: ${product['thumbnail']}');
-      //   print('Product Images: ${product['images']}');
-      //   print('');        
-      // }
-      
-      // return productFromJson(prods);
-      // return data.map((e) => productFromJson(e)).toList();
-      // if (response.statusCode == 200) {
-      //   var prod = jsonDecode(response.body[1]);
-      //   return productFromJson(prod);
-      // }      
+      return productList;      
     } catch (e) {
       log(e.toString());
       return null;
